+++ csrc/gpu/aten/operators/xetla/mha.h
@@ -24,8 +24,8 @@ struct fmha_forward_kernel_args_t {
   float alpha;
   float beta;
   float dropout_prob;
-  int64_t* cu_seqlen_q;
-  int64_t* cu_seqlen_k;
+  int32_t* cu_seqlen_q;
+  int32_t* cu_seqlen_k;
   uint32_t num_batches;
   uint32_t num_heads;
   uint32_t num_kv_heads;
